+++ airbyte-webapp/src/pages/connections/ConnectionTransformationPage/DbtCloudTransformationsCard/DbtCloudTransformationsCard.module.scss
@@ -1,0 +1,1 @@
+@forward "./DbtCloudCard.scss";
