+++ airbyte-webapp/src/pages/connections/ConnectionTransformationPage/DbtCloudTransformationsCard/DbtJobsForm.module.scss
@@ -1,0 +1,5 @@
+@forward "./DbtCloudCard.scss";
+
+.jobListForm {
+  width: 100%;
+}
