+++ airbyte-webapp/src/pages/connections/ConnectionTransformationPage/DbtCloudTransformationsCard/JobsListItem.module.scss
@@ -1,49 +1,6 @@
 @use "scss/colors";
 @use "scss/variables";
 
-.jobListContainer {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  padding: variables.$spacing-xl;
-  background-color: colors.$grey-50;
-}
-
-.jobListTitle {
-  display: flex;
-  justify-content: space-between;
-}
-
-.jobListForm {
-  width: 100%;
-}
-
-.emptyListImage {
-  width: 111px;
-  height: 111px;
-  margin: variables.$spacing-xl 0;
-}
-
-.contextExplanation {
-  color: colors.$grey-300;
-  width: 100%;
-
-  & a {
-    color: colors.$grey-300;
-  }
-}
-
-.jobListButtonGroup {
-  display: flex;
-  justify-content: flex-end;
-  margin-top: variables.$spacing-xl;
-  width: 100%;
-}
-
-.jobListButton {
-  margin-left: variables.$spacing-md;
-}
-
 .jobListItem {
   margin-top: variables.$spacing-md;
   padding: variables.$spacing-md variables.$spacing-xl;
@@ -53,16 +10,16 @@
   align-items: center;
 }
 
-.dbtLogo {
-  height: 18px;
-  width: 18px;
-  margin-right: variables.$spacing-md;
-}
-
 .jobListItemIntegrationName {
   display: flex;
   align-items: center;
   flex: 1 2 auto;
+}
+
+.dbtLogo {
+  height: 18px;
+  width: 18px;
+  margin-right: variables.$spacing-md;
 }
 
 .jobListItemIdFieldGroup {
