+++ airbyte-webapp/src/pages/connections/ConnectionTransformationPage/DbtCloudTransformationsCard/NoDbtIntegration.module.scss
@@ -1,0 +1,1 @@
+@forward "./DbtCloudCard.scss";
